backend/pyproject.toml
[project]
name = "pscweb3"
version = "0.1.0"
description = "演劇制作管理システム - FastAPI + Azure Functions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "pydantic>=2.0",
    "pydantic-settings",
    "authlib",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "httpx",
    "azure-functions",
    "fountain",
    "alembic",
    "playscript",
    "structlog>=25.5.0",
    "itsdangerous>=2.2.0",
    "requests>=2.32.5",
    "pynacl>=1.6.1",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "aiosqlite",
    "httpx",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
fountain = { git = "https://github.com/Tagirijus/fountain.git", rev = "7da5447abae640f34448dd36fee83f47a7415fcf" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "httpx>=0.28.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.8",
    "ty>=0.0.1a32",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations
    "D",      # pydocstyle
]
ignore = [
    # Type annotation
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
    
    # Format compatibility
    "COM812",  # trailing-comma-missing
    "ISC001",  # single-line-implicit-string-concatenation
    
    # Docstring formatting (conflicts or Japanese compatibility)
    "D100",    # undocumented-public-module
    "D104",    # undocumented-public-package
    "D200",    # fits-on-one-line
    "D203",    # one-blank-line-before-class
    "D205",    # blank-line-after-summary  
    "D207",    # docstring-under-indented
    "D208",    # docstring-over-indented
    "D209",    # newline-before-docstring
    "D212",    # multi-line-summary-first-line
    "D213",    # multi-line-summary-second-line
    
    # Docstring content (Japanese compatibility)
    "D400",    # ends-in-period
    "D401",    # non-imperative-mood
    "D404",    # docstring-starts-with-this
    "D415",    # ends-in-punctuation
    "D417",    # undocumented-param
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "D"]
"alembic/*" = ["ANN", "D"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
